optimize image resize function

+++ routes/userIcon.js
@@ -7,6 +7,7 @@ var generalFunc=require('./express_component/generalFunction').generateFunction
 var general=require('./assist/general').general
 var runtimeNodeError=require('./error_define/runtime_node_error').runtime_node_error
 var fs=require('fs')
+//var generalFunc=require('./express_component/generalFunction').generateFunction
 
 var multiparty = require('multiparty');
 var global=require('./error_define/global').global
@@ -20,7 +21,7 @@ router.get('/',function(req,res,next){
     if(preResult.rc>0){
         return res.json(preResult)
     }
-    res.render('userIcon',{title:'首页',year:new Date().getFullYear()});
+    res.render('userIcon',{title:'首页',year:new Date().getFullYear(),userIconWidth:global.userIconWidth.define,userIconHeight:global.userIconHeight.define});
 })
 
 router.post('/userIcon/upload',function(res,req,next){
@@ -28,12 +29,21 @@ router.post('/userIcon/upload',function(res,req,next){
 })
 
 router.post('/upload',function(req,res,next){
-    //console.log('upload')
-    var form = new multiparty.Form({uploadDir:global.userIconUploadDir.define ,maxFilesSize:global.userIconMaxSizeMultiParty.define});
+    //解析以GM格式定义的文件大小，转换成byte，传递给MultiParty
+    var definedImageFileSize=global.userIconMaxSizeGm.define;
+    var parseDefinedImageFileSize=generalFunc.parseGmFileSize(definedImageFileSize)
+    if(parseDefinedImageFileSize.rc>0){
+        return res.json(parseDefinedImageFileSize)
+    }
+    //console.log(parseDefinedImageFileSize)
+    //根据解析的结果，转换成byte为单位的数值
+    var multiPartyMaxSize=generalFunc.convertImageFileSizeToByte(parseDefinedImageFileSize.msg.sizeNum,parseDefinedImageFileSize.msg.sizeUnit)
+//console.log(multiPartyMaxSize)
+    var form = new multiparty.Form({uploadDir:global.userIconUploadDir.define ,maxFilesSize:multiPartyMaxSize.msg});
     //console.log(form)
     form.parse(req, function (err, fields, files) {
         if(err){
-            return res.json(global.userIconMaxSizeMultiParty.client)
+            return res.json(global.userIconMaxSizeGm.client)
         }
 /*        var filesTmp = JSON.stringify(files, null, 2);
         var fieldsTemp = JSON.stringify(fields, null, 2);*/
@@ -75,7 +85,7 @@ router.post('/upload',function(req,res,next){
                     }
 /*console.log(result)
 console.log(result1)*/
-                    var p=/(\d{1,}\.?\d{1,})([KkMmGg]i)?/ //1.8Ki
+/*                    var p=/(\d{1,}\.?\d{1,})([KkMmGg]i)?/ //1.8Ki
                     var fileSize=fileSizeResult.msg.match(p)
                         //[ '1.812ki', '1.812', 'ki', index: 0, input: '1.812ki' ]
                     //unit可能为空，因为文件小于1k
@@ -85,12 +95,21 @@ console.log(result1)*/
                     var fileSizeNum=parseFloat(fileSize[1])
                     if(isNaN(fileSizeNum)){
                         callback({},runtimeNodeError.image.cantParseFileSizeNum)
+                    }*/
+                    //解析实际文件大小（GM）。实际并不需要，MultiParty已经保证了不会超出
+                    var uploadFileSize=generalFunc.parseGmFileSize(fileSizeResult.msg)
+                    if(uploadFileSize.rc>0){
+                        return res.json(uploadFileSize)
                     }
-                    var fileSizeUnit=fileSize[2]
-                    if('Mi'==fileSizeUnit || 'Gi'==fileSize || fileSizeNum>global.userIconMaxSizeGm.define){
+                    var uploadFileSizeInByte=generalFunc.convertImageFileSizeToByte(uploadFileSize.msg.sizeNum,uploadFileSize.msg.sizeUnit)
+                    if(uploadFileSizeInByte.rc>0){
+                        return res.json(uploadFileSizeInByte)
+                    }
+//console.log(uploadFileSizeInByte.msg)
+                    if(uploadFileSizeInByte.msg>global.userIconMaxSizeGm.define){
                         callback({},global.userIconMaxSizeGm.client)
                     }
-                    callback(null,fileTypeResult,file)
+                    callback(null,fileTypeResult)
                 })
 
             },
@@ -108,23 +127,45 @@ console.log(result1)*/
                     //生成hash名
                     var fileType=fileTypeResult.msg.toLowerCase();
                     var hashName=hash.hash(file.originalFilename+new Date().getTime(),'sha1')
+
+                    var destFilePath=global.userIconUploadDir.define+hashName +'.'+fileType
+                    //console.log(fileWHResult)
+                    //console.log(destFilePath)
+                    //console.log(fileWHResult.msg.width)
+                    //console.log(fileWHResult.msg.height)
+                    callback(null,fileWHResult.msg.width,fileWHResult.msg.height,destFilePath)
+                    //console.log(callback)
+                    //callback(null,1366,768,destFilePath)
                     //if(fileWHResult.msg.width>global.userIconWidth || fileWHResult.msg.height>global.userIconHeight){
                     //console.log(filePath)
                     //console.log(global.userIconUploadDir.define+hashName+'.'+fileType)
-                        image.command.resizeUserIcon(filePath,global.userIconUploadDir.define+hashName+'.'+fileType,global.userIconWidth.define,global.userIconHeight.define,function(err,result){
-                            if(err){
-console.log(err)
-                                callback(err,result)
-                            }
-console.log(result)
-                            callback(null,result)
-                        })
+
                     //}
                 })
 
+            },
+        //    4. 根据width决定如何处理：直接改名还是resize
+            function(fileWidth,fileHeight,destFilePath,callback){
+
+                if(fileWidth===global.userIconWidth.define && fileHeight===global.userIconHeight.define){
+                    fs.renameSync(filePath,destFilePath)
+                    callback(null)
+                }else{
+                    image.command.resizeUserIcon(filePath,destFilePath,global.userIconWidth.define,global.userIconHeight.define,function(err,result){
+                        if(err){
+                            //console.log(err)
+                            callback(err,result)
+                        }
+                        //console.log(result)
+                        fs.unlinkSync(filePath)
+                        callback(null,result)
+                    })
+                }
             }
         ],function(err,result){
-            fs.unlinkSync(filePath)
+            //if(err){
+                return res.json(result)
+            //}
         })
         //if()
